Move command functions to user command util for reusing in login
