+++ src/commands/user/signup.ts
@@ -1,61 +1,10 @@
 import chalk from 'chalk';
-import inquirer from 'inquirer';
 
 import { controller } from '../../api';
 import * as config from '../../config';
 import { Arguments, createCommand } from '../../util';
 import requireClusterConfig from '../middleware/requireClusterConfig';
-import { updateGlobalConfigUser } from './util';
-
-const emailPrompt: inquirer.Question = {
-  message: 'Email:',
-  name: 'email',
-  type: 'input',
-  validate: (input) => {
-    if (input.trim() === '') {
-    return 'Email is required.';
-    }
-    if (input.indexOf('@') === -1) {
-    return 'Email is not valid (must contains @).';
-    }
-    return true;
-  }
-};
-
-const passwordPrompt: inquirer.Question = {
-  message: 'Password:',
-  name: 'password',
-  type: 'password',
-  validate: (input) => {
-    if (input === '') {
-    return 'Password is required.';
-    }
-    return true;
-  }
-};
-
-function askCredentials(argv: Arguments) {
-  const prompts = [];
-  const credentials = {
-    email: argv.email as string,
-    password: argv.password as string
-  };
-
-  if (credentials.email) {
-    console.log(`Sign up as ${credentials.email}.`);
-  } else {
-    prompts.push(emailPrompt);
-  }
-
-  prompts.push(passwordPrompt);
-
-  return inquirer.prompt(prompts).then((answers) => {
-    return {
-      ...credentials,
-      ...answers
-    };
-  });
-}
+import { askCredentials, updateGlobalConfigUser } from './util';
 
 function run(argv: Arguments) {
   let email: string;
@@ -83,9 +32,13 @@ function run(argv: Arguments) {
 export default createCommand({
   builder: (yargs) => {
     return yargs
-      .middleware(requireClusterConfig);
+      .middleware(requireClusterConfig)
+      .option('email', {
+        desc: 'Sign up as email',
+        type: 'string'
+      });
   },
   command: 'signup',
-  describe: 'Sign up skygear cluster user',
+  describe: 'Sign up Skygear cluster user',
   handler: run
 });
